+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- The MP's name is now their full name with title (e.g. The Right Honourable)
+
 ## [Release-89][release-89]
 
 ### Changed